% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_core_housing_needs_tables.R
\name{prepare_core_housing_needs_tables}
\alias{prepare_core_housing_needs_tables}
\title{Combine CHN tables}
\usage{
prepare_core_housing_needs_tables(
  input_path,
  output_path = NULL,
  write_output = TRUE
)
}
\arguments{
\item{input_path}{Path to the input CSV file.}

\item{output_path}{Path to save the output CSV file. If NULL, the file will be
saved in the same directory as the input file with the name
"combined_table.csv".}

\item{write_output}{Logical. Should the combined table be written to a file?
Default is TRUE.}
}
\value{
A data frame containing the combined data.
}
\description{
This function reads in a table of data on core housing needs by municipality
and creates a combined table that separates out each variable in the data into
its own column. The resulting table is exported as a CSV file if write_output
is TRUE.
}
